Currencies init from constants, rename dt.date to date, remove types from docstrings
