+++ pycbrf/constants.py
@@ -1,0 +1,5 @@
+URL_BASE = 'http://www.cbr.ru/scripts/'
+
+DAILY_CURRENCIES = b'<?xml version="1.0" encoding="windows-1251"?><Valuta name="Foreign Currency Market Lib"><Item ID="R01010"><Name>\xc0\xe2\xf1\xf2\xf0\xe0\xeb\xe8\xe9\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Australian Dollar</EngName><Nominal>1</Nominal><ParentCode>R01010    </ParentCode><ISO_Num_Code>36</ISO_Num_Code><ISO_Char_Code>AUD</ISO_Char_Code></Item><Item ID="R01015"><Name>\xc0\xe2\xf1\xf2\xf0\xe8\xe9\xf1\xea\xe8\xe9 \xf8\xe8\xeb\xeb\xe8\xed\xe3</Name><EngName>Austrian Shilling</EngName><Nominal>1000</Nominal><ParentCode>R01015    </ParentCode><ISO_Num_Code>40</ISO_Num_Code><ISO_Char_Code>ATS</ISO_Char_Code></Item><Item ID="R01020A"><Name>\xc0\xe7\xe5\xf0\xe1\xe0\xe9\xe4\xe6\xe0\xed\xf1\xea\xe8\xe9 \xec\xe0\xed\xe0\xf2</Name><EngName>Azerbaijan Manat</EngName><Nominal>1</Nominal><ParentCode>R01020    </ParentCode><ISO_Num_Code>944</ISO_Num_Code><ISO_Char_Code>AZN</ISO_Char_Code></Item><Item ID="R01035"><Name>\xd4\xf3\xed\xf2 \xf1\xf2\xe5\xf0\xeb\xe8\xed\xe3\xee\xe2 \xd1\xee\xe5\xe4\xe8\xed\xe5\xed\xed\xee\xe3\xee \xea\xee\xf0\xee\xeb\xe5\xe2\xf1\xf2\xe2\xe0</Name><EngName>British Pound Sterling</EngName><Nominal>1</Nominal><ParentCode>R01035    </ParentCode><ISO_Num_Code>826</ISO_Num_Code><ISO_Char_Code>GBP</ISO_Char_Code></Item><Item ID="R01040F"><Name>\xc0\xed\xe3\xee\xeb\xfc\xf1\xea\xe0\xff \xed\xee\xe2\xe0\xff \xea\xe2\xe0\xed\xe7\xe0</Name><EngName>Angolan new Kwanza</EngName><Nominal>100000</Nominal><ParentCode>R01040    </ParentCode><ISO_Num_Code>24</ISO_Num_Code><ISO_Char_Code>AON</ISO_Char_Code></Item><Item ID="R01060"><Name>\xc0\xf0\xec\xff\xed\xf1\xea\xe8\xe9 \xe4\xf0\xe0\xec</Name><EngName>Armenia Dram</EngName><Nominal>1000</Nominal><ParentCode>R01060    </ParentCode><ISO_Num_Code>51</ISO_Num_Code><ISO_Char_Code>AMD</ISO_Char_Code></Item><Item ID="R01090B"><Name>\xc1\xe5\xeb\xee\xf0\xf3\xf1\xf1\xea\xe8\xe9 \xf0\xf3\xe1\xeb\xfc</Name><EngName>Belarussian Ruble</EngName><Nominal>1</Nominal><ParentCode>R01090    </ParentCode><ISO_Num_Code>933</ISO_Num_Code><ISO_Char_Code>BYN</ISO_Char_Code></Item><Item ID="R01095"><Name>\xc1\xe5\xeb\xfc\xe3\xe8\xe9\xf1\xea\xe8\xe9 \xf4\xf0\xe0\xed\xea</Name><EngName>Belgium Franc</EngName><Nominal>1000</Nominal><ParentCode>R01095    </ParentCode><ISO_Num_Code>56</ISO_Num_Code><ISO_Char_Code>BEF</ISO_Char_Code></Item><Item ID="R01100"><Name>\xc1\xee\xeb\xe3\xe0\xf0\xf1\xea\xe8\xe9 \xeb\xe5\xe2</Name><EngName>Bulgarian lev</EngName><Nominal>1</Nominal><ParentCode>R01100    </ParentCode><ISO_Num_Code>975</ISO_Num_Code><ISO_Char_Code>BGN</ISO_Char_Code></Item><Item ID="R01115"><Name>\xc1\xf0\xe0\xe7\xe8\xeb\xfc\xf1\xea\xe8\xe9 \xf0\xe5\xe0\xeb</Name><EngName>Brazil Real</EngName><Nominal>1</Nominal><ParentCode>R01115    </ParentCode><ISO_Num_Code>986</ISO_Num_Code><ISO_Char_Code>BRL</ISO_Char_Code></Item><Item ID="R01135"><Name>\xc2\xe5\xed\xe3\xe5\xf0\xf1\xea\xe8\xe9 \xf4\xee\xf0\xe8\xed\xf2</Name><EngName>Hungarian Forint</EngName><Nominal>100</Nominal><ParentCode>R01135    </ParentCode><ISO_Num_Code>348</ISO_Num_Code><ISO_Char_Code>HUF</ISO_Char_Code></Item><Item ID="R01200"><Name>\xc3\xee\xed\xea\xee\xed\xe3\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Hong Kong Dollar</EngName><Nominal>10</Nominal><ParentCode>R01200    </ParentCode><ISO_Num_Code>344</ISO_Num_Code><ISO_Char_Code>HKD</ISO_Char_Code></Item><Item ID="R01205"><Name>\xc3\xf0\xe5\xf7\xe5\xf1\xea\xe0\xff \xe4\xf0\xe0\xf5\xec\xe0</Name><EngName>Greek Drachma</EngName><Nominal>10000</Nominal><ParentCode>R01205    </ParentCode><ISO_Num_Code>300</ISO_Num_Code><ISO_Char_Code>GRD</ISO_Char_Code></Item><Item ID="R01215"><Name>\xc4\xe0\xf2\xf1\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Danish Krone</EngName><Nominal>10</Nominal><ParentCode>R01215    </ParentCode><ISO_Num_Code>208</ISO_Num_Code><ISO_Char_Code>DKK</ISO_Char_Code></Item><Item ID="R01235"><Name>\xc4\xee\xeb\xeb\xe0\xf0 \xd1\xd8\xc0</Name><EngName>US Dollar</EngName><Nominal>1</Nominal><ParentCode>R01235    </ParentCode><ISO_Num_Code>840</ISO_Num_Code><ISO_Char_Code>USD</ISO_Char_Code></Item><Item ID="R01239"><Name>\xc5\xe2\xf0\xee</Name><EngName>Euro</EngName><Nominal>1</Nominal><ParentCode>R01239    </ParentCode><ISO_Num_Code>978</ISO_Num_Code><ISO_Char_Code>EUR</ISO_Char_Code></Item><Item ID="R01270"><Name>\xc8\xed\xe4\xe8\xe9\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Indian Rupee</EngName><Nominal>100</Nominal><ParentCode>R01270    </ParentCode><ISO_Num_Code>356</ISO_Num_Code><ISO_Char_Code>INR</ISO_Char_Code></Item><Item ID="R01305"><Name>\xc8\xf0\xeb\xe0\xed\xe4\xf1\xea\xe8\xe9 \xf4\xf3\xed\xf2</Name><EngName>Irish Pound</EngName><Nominal>100</Nominal><ParentCode>R01305    </ParentCode><ISO_Num_Code>372</ISO_Num_Code><ISO_Char_Code>IEP</ISO_Char_Code></Item><Item ID="R01310"><Name>\xc8\xf1\xeb\xe0\xed\xe4\xf1\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Iceland Krona</EngName><Nominal>10000</Nominal><ParentCode>R01310    </ParentCode><ISO_Num_Code>352</ISO_Num_Code><ISO_Char_Code>ISK</ISO_Char_Code></Item><Item ID="R01315"><Name>\xc8\xf1\xef\xe0\xed\xf1\xea\xe0\xff \xef\xe5\xf1\xe5\xf2\xe0</Name><EngName>Spanish Peseta</EngName><Nominal>10000</Nominal><ParentCode>R01315    </ParentCode><ISO_Num_Code>724</ISO_Num_Code><ISO_Char_Code>ESP</ISO_Char_Code></Item><Item ID="R01325"><Name>\xc8\xf2\xe0\xeb\xfc\xff\xed\xf1\xea\xe0\xff \xeb\xe8\xf0\xe0</Name><EngName>Italian Lira</EngName><Nominal>100000</Nominal><ParentCode>R01325    </ParentCode><ISO_Num_Code>380</ISO_Num_Code><ISO_Char_Code>ITL</ISO_Char_Code></Item><Item ID="R01335"><Name>\xca\xe0\xe7\xe0\xf5\xf1\xf2\xe0\xed\xf1\xea\xe8\xe9 \xf2\xe5\xed\xe3\xe5</Name><EngName>Kazakhstan Tenge</EngName><Nominal>100</Nominal><ParentCode>R01335    </ParentCode><ISO_Num_Code>398</ISO_Num_Code><ISO_Char_Code>KZT</ISO_Char_Code></Item><Item ID="R01350"><Name>\xca\xe0\xed\xe0\xe4\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Canadian Dollar</EngName><Nominal>1</Nominal><ParentCode>R01350    </ParentCode><ISO_Num_Code>124</ISO_Num_Code><ISO_Char_Code>CAD</ISO_Char_Code></Item><Item ID="R01370"><Name>\xca\xe8\xf0\xe3\xe8\xe7\xf1\xea\xe8\xe9 \xf1\xee\xec</Name><EngName>Kyrgyzstan Som</EngName><Nominal>100</Nominal><ParentCode>R01370    </ParentCode><ISO_Num_Code>417</ISO_Num_Code><ISO_Char_Code>KGS</ISO_Char_Code></Item><Item ID="R01375"><Name>\xca\xe8\xf2\xe0\xe9\xf1\xea\xe8\xe9 \xfe\xe0\xed\xfc</Name><EngName>China Yuan</EngName><Nominal>10</Nominal><ParentCode>R01375    </ParentCode><ISO_Num_Code>156</ISO_Num_Code><ISO_Char_Code>CNY</ISO_Char_Code></Item><Item ID="R01390"><Name>\xca\xf3\xe2\xe5\xe9\xf2\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Kuwaiti Dinar</EngName><Nominal>10</Nominal><ParentCode>R01390    </ParentCode><ISO_Num_Code>414</ISO_Num_Code><ISO_Char_Code>KWD</ISO_Char_Code></Item><Item ID="R01405"><Name>\xcb\xe0\xf2\xe2\xe8\xe9\xf1\xea\xe8\xe9 \xeb\xe0\xf2</Name><EngName>Latvian Lat</EngName><Nominal>1</Nominal><ParentCode>R01405    </ParentCode><ISO_Num_Code>428</ISO_Num_Code><ISO_Char_Code>LVL</ISO_Char_Code></Item><Item ID="R01420"><Name>\xcb\xe8\xe2\xe0\xed\xf1\xea\xe8\xe9 \xf4\xf3\xed\xf2</Name><EngName>Lebanese Pound</EngName><Nominal>100000</Nominal><ParentCode>R01420    </ParentCode><ISO_Num_Code>422</ISO_Num_Code><ISO_Char_Code>LBP</ISO_Char_Code></Item><Item ID="R01435"><Name>\xcb\xe8\xf2\xee\xe2\xf1\xea\xe8\xe9 \xeb\xe8\xf2</Name><EngName>Lithuanian Lita</EngName><Nominal>1</Nominal><ParentCode>R01435    </ParentCode><ISO_Num_Code>440</ISO_Num_Code><ISO_Char_Code>LTL</ISO_Char_Code></Item><Item ID="R01436"><Name>\xcb\xe8\xf2\xee\xe2\xf1\xea\xe8\xe9 \xf2\xe0\xeb\xee\xed</Name><EngName>Lithuanian talon</EngName><Nominal>1</Nominal><ParentCode>R01435    </ParentCode><ISO_Num_Code></ISO_Num_Code><ISO_Char_Code></ISO_Char_Code></Item><Item ID="R01500"><Name>\xcc\xee\xeb\xe4\xe0\xe2\xf1\xea\xe8\xe9 \xeb\xe5\xe9</Name><EngName>Moldova Lei</EngName><Nominal>10</Nominal><ParentCode>R01500    </ParentCode><ISO_Num_Code>498</ISO_Num_Code><ISO_Char_Code>MDL</ISO_Char_Code></Item><Item ID="R01510"><Name>\xcd\xe5\xec\xe5\xf6\xea\xe0\xff \xec\xe0\xf0\xea\xe0</Name><EngName>Deutsche Mark</EngName><Nominal>1</Nominal><ParentCode>R01510    </ParentCode><ISO_Num_Code>276</ISO_Num_Code><ISO_Char_Code>DEM</ISO_Char_Code></Item><Item ID="R01510A"><Name>\xcd\xe5\xec\xe5\xf6\xea\xe0\xff \xec\xe0\xf0\xea\xe0</Name><EngName>Deutsche Mark</EngName><Nominal>100</Nominal><ParentCode>R01510    </ParentCode><ISO_Num_Code>280</ISO_Num_Code><ISO_Char_Code>DEM</ISO_Char_Code></Item><Item ID="R01523"><Name>\xcd\xe8\xe4\xe5\xf0\xeb\xe0\xed\xe4\xf1\xea\xe8\xe9 \xe3\xf3\xeb\xfc\xe4\xe5\xed</Name><EngName>Netherlands Gulden</EngName><Nominal>100</Nominal><ParentCode>R01523    </ParentCode><ISO_Num_Code>528</ISO_Num_Code><ISO_Char_Code>NLG</ISO_Char_Code></Item><Item ID="R01535"><Name>\xcd\xee\xf0\xe2\xe5\xe6\xf1\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Norwegian Krone</EngName><Nominal>10</Nominal><ParentCode>R01535    </ParentCode><ISO_Num_Code>578</ISO_Num_Code><ISO_Char_Code>NOK</ISO_Char_Code></Item><Item ID="R01565"><Name>\xcf\xee\xeb\xfc\xf1\xea\xe8\xe9 \xe7\xeb\xee\xf2\xfb\xe9</Name><EngName>Polish Zloty</EngName><Nominal>1</Nominal><ParentCode>R01565    </ParentCode><ISO_Num_Code>985</ISO_Num_Code><ISO_Char_Code>PLN</ISO_Char_Code></Item><Item ID="R01570"><Name>\xcf\xee\xf0\xf2\xf3\xe3\xe0\xeb\xfc\xf1\xea\xe8\xe9 \xfd\xf1\xea\xf3\xe4\xee</Name><EngName>Portuguese Escudo</EngName><Nominal>10000</Nominal><ParentCode>R01570    </ParentCode><ISO_Num_Code>620</ISO_Num_Code><ISO_Char_Code>PTE</ISO_Char_Code></Item><Item ID="R01585"><Name>\xd0\xf3\xec\xfb\xed\xf1\xea\xe8\xe9 \xeb\xe5\xe9</Name><EngName>Romanian Leu</EngName><Nominal>10000</Nominal><ParentCode>R01585    </ParentCode><ISO_Num_Code>642</ISO_Num_Code><ISO_Char_Code>ROL</ISO_Char_Code></Item><Item ID="R01585F"><Name>\xd0\xf3\xec\xfb\xed\xf1\xea\xe8\xe9 \xeb\xe5\xe9</Name><EngName>Romanian Leu</EngName><Nominal>10</Nominal><ParentCode>R01585    </ParentCode><ISO_Num_Code>946</ISO_Num_Code><ISO_Char_Code>RON</ISO_Char_Code></Item><Item ID="R01589"><Name>\xd1\xc4\xd0 (\xf1\xef\xe5\xf6\xe8\xe0\xeb\xfc\xed\xfb\xe5 \xef\xf0\xe0\xe2\xe0 \xe7\xe0\xe8\xec\xf1\xf2\xe2\xee\xe2\xe0\xed\xe8\xff)</Name><EngName>SDR</EngName><Nominal>1</Nominal><ParentCode>R01589    </ParentCode><ISO_Num_Code>960</ISO_Num_Code><ISO_Char_Code>XDR</ISO_Char_Code></Item><Item ID="R01625"><Name>\xd1\xe8\xed\xe3\xe0\xef\xf3\xf0\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Singapore Dollar</EngName><Nominal>1</Nominal><ParentCode>R01625    </ParentCode><ISO_Num_Code>702</ISO_Num_Code><ISO_Char_Code>SGD</ISO_Char_Code></Item><Item ID="R01665A"><Name>\xd1\xf3\xf0\xe8\xed\xe0\xec\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Surinam Dollar</EngName><Nominal>1</Nominal><ParentCode>R01665    </ParentCode><ISO_Num_Code>968</ISO_Num_Code><ISO_Char_Code>SRD</ISO_Char_Code></Item><Item ID="R01670"><Name>\xd2\xe0\xe4\xe6\xe8\xea\xf1\xea\xe8\xe9 \xf1\xee\xec\xee\xed\xe8</Name><EngName>Tajikistan Ruble</EngName><Nominal>10</Nominal><ParentCode>R01670    </ParentCode><ISO_Num_Code>972</ISO_Num_Code><ISO_Char_Code>TJS</ISO_Char_Code></Item><Item ID="R01670B"><Name>\xd2\xe0\xe4\xe6\xe8\xea\xf1\xea\xe8\xe9 \xf0\xf3\xe1\xeb</Name><EngName>Tajikistan Ruble</EngName><Nominal>10</Nominal><ParentCode>R01670    </ParentCode><ISO_Num_Code>762</ISO_Num_Code><ISO_Char_Code>TJR</ISO_Char_Code></Item><Item ID="R01700J"><Name>\xd2\xf3\xf0\xe5\xf6\xea\xe0\xff \xeb\xe8\xf0\xe0</Name><EngName>Turkish Lira</EngName><Nominal>1</Nominal><ParentCode>R01700    </ParentCode><ISO_Num_Code>949</ISO_Num_Code><ISO_Char_Code>TRY</ISO_Char_Code></Item><Item ID="R01710"><Name>\xd2\xf3\xf0\xea\xec\xe5\xed\xf1\xea\xe8\xe9 \xec\xe0\xed\xe0\xf2</Name><EngName>Turkmenistan Manat</EngName><Nominal>10000</Nominal><ParentCode>R01710    </ParentCode><ISO_Num_Code>795</ISO_Num_Code><ISO_Char_Code>TMM</ISO_Char_Code></Item><Item ID="R01710A"><Name>\xcd\xee\xe2\xfb\xe9 \xf2\xf3\xf0\xea\xec\xe5\xed\xf1\xea\xe8\xe9 \xec\xe0\xed\xe0\xf2</Name><EngName>New Turkmenistan Manat</EngName><Nominal>1</Nominal><ParentCode>R01710    </ParentCode><ISO_Num_Code>934</ISO_Num_Code><ISO_Char_Code>TMT</ISO_Char_Code></Item><Item ID="R01717"><Name>\xd3\xe7\xe1\xe5\xea\xf1\xea\xe8\xe9 \xf1\xf3\xec</Name><EngName>Uzbekistan Sum</EngName><Nominal>1000</Nominal><ParentCode>R01717    </ParentCode><ISO_Num_Code>860</ISO_Num_Code><ISO_Char_Code>UZS</ISO_Char_Code></Item><Item ID="R01720"><Name>\xd3\xea\xf0\xe0\xe8\xed\xf1\xea\xe0\xff \xe3\xf0\xe8\xe2\xed\xe0</Name><EngName>Ukrainian Hryvnia</EngName><Nominal>10</Nominal><ParentCode>R01720    </ParentCode><ISO_Num_Code>980</ISO_Num_Code><ISO_Char_Code>UAH</ISO_Char_Code></Item><Item ID="R01720A"><Name>\xd3\xea\xf0\xe0\xe8\xed\xf1\xea\xe8\xe9 \xea\xe0\xf0\xe1\xee\xe2\xe0\xed\xe5\xf6</Name><EngName>Ukrainian Hryvnia</EngName><Nominal>1</Nominal><ParentCode>R01720    </ParentCode><ISO_Num_Code></ISO_Num_Code><ISO_Char_Code></ISO_Char_Code></Item><Item ID="R01740"><Name>\xd4\xe8\xed\xeb\xff\xed\xe4\xf1\xea\xe0\xff \xec\xe0\xf0\xea\xe0</Name><EngName>Finnish Marka</EngName><Nominal>100</Nominal><ParentCode>R01740    </ParentCode><ISO_Num_Code>246</ISO_Num_Code><ISO_Char_Code>FIM</ISO_Char_Code></Item><Item ID="R01750"><Name>\xd4\xf0\xe0\xed\xf6\xf3\xe7\xf1\xea\xe8\xe9 \xf4\xf0\xe0\xed\xea</Name><EngName>French Franc</EngName><Nominal>1000</Nominal><ParentCode>R01750    </ParentCode><ISO_Num_Code>250</ISO_Num_Code><ISO_Char_Code>FRF</ISO_Char_Code></Item><Item ID="R01760"><Name>\xd7\xe5\xf8\xf1\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Czech Koruna</EngName><Nominal>10</Nominal><ParentCode>R01760    </ParentCode><ISO_Num_Code>203</ISO_Num_Code><ISO_Char_Code>CZK</ISO_Char_Code></Item><Item ID="R01770"><Name>\xd8\xe2\xe5\xe4\xf1\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Swedish Krona</EngName><Nominal>10</Nominal><ParentCode>R01770    </ParentCode><ISO_Num_Code>752</ISO_Num_Code><ISO_Char_Code>SEK</ISO_Char_Code></Item><Item ID="R01775"><Name>\xd8\xe2\xe5\xe9\xf6\xe0\xf0\xf1\xea\xe8\xe9 \xf4\xf0\xe0\xed\xea</Name><EngName>Swiss Franc</EngName><Nominal>1</Nominal><ParentCode>R01775    </ParentCode><ISO_Num_Code>756</ISO_Num_Code><ISO_Char_Code>CHF</ISO_Char_Code></Item><Item ID="R01790"><Name>\xdd\xca\xde</Name><EngName>ECU</EngName><Nominal>1</Nominal><ParentCode>R01790    </ParentCode><ISO_Num_Code>954</ISO_Num_Code><ISO_Char_Code>XEU</ISO_Char_Code></Item><Item ID="R01795"><Name>\xdd\xf1\xf2\xee\xed\xf1\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Estonian Kroon</EngName><Nominal>10</Nominal><ParentCode>R01795    </ParentCode><ISO_Num_Code>233</ISO_Num_Code><ISO_Char_Code>EEK</ISO_Char_Code></Item><Item ID="R01805"><Name>\xde\xe3\xee\xf1\xeb\xe0\xe2\xf1\xea\xe8\xe9 \xed\xee\xe2\xfb\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Yugoslavian Dinar</EngName><Nominal>1</Nominal><ParentCode>R01804    </ParentCode><ISO_Num_Code>890</ISO_Num_Code><ISO_Char_Code>YUN</ISO_Char_Code></Item><Item ID="R01810"><Name>\xde\xe6\xed\xee\xe0\xf4\xf0\xe8\xea\xe0\xed\xf1\xea\xe8\xe9 \xf0\xfd\xed\xe4</Name><EngName>S.African Rand</EngName><Nominal>10</Nominal><ParentCode>R01810    </ParentCode><ISO_Num_Code>710</ISO_Num_Code><ISO_Char_Code>ZAR</ISO_Char_Code></Item><Item ID="R01815"><Name>\xc2\xee\xed \xd0\xe5\xf1\xef\xf3\xe1\xeb\xe8\xea\xe8 \xca\xee\xf0\xe5\xff</Name><EngName>South Korean Won</EngName><Nominal>1000</Nominal><ParentCode>R01815    </ParentCode><ISO_Num_Code>410</ISO_Num_Code><ISO_Char_Code>KRW</ISO_Char_Code></Item><Item ID="R01820"><Name>\xdf\xef\xee\xed\xf1\xea\xe0\xff \xe8\xe5\xed\xe0</Name><EngName>Japanese Yen</EngName><Nominal>100</Nominal><ParentCode>R01820    </ParentCode><ISO_Num_Code>392</ISO_Num_Code><ISO_Char_Code>JPY</ISO_Char_Code></Item></Valuta>'
+
+MONTHLY_CURRENCIES = b'<?xml version="1.0" encoding="windows-1251"?><Valuta name="Foreign Currency Market Lib"><Item ID="R01025"><Name>\xc0\xeb\xe1\xe0\xed\xf1\xea\xe8\xe9 \xeb\xe5\xea</Name><EngName>Albanian Lek</EngName><Nominal>100</Nominal><ParentCode>R01025    </ParentCode><ISO_Num_Code>8</ISO_Num_Code><ISO_Char_Code>ALL</ISO_Char_Code></Item><Item ID="R01030"><Name>\xc0\xeb\xe6\xe8\xf0\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Algerian Dinar</EngName><Nominal>100</Nominal><ParentCode>R01030    </ParentCode><ISO_Num_Code>12</ISO_Num_Code><ISO_Char_Code>DZD</ISO_Char_Code></Item><Item ID="R01040"><Name>\xc0\xed\xe3\xee\xeb\xfc\xf1\xea\xe0\xff \xea\xe2\xe0\xed\xe7\xe0</Name><EngName>Angolan Kwanza</EngName><Nominal>100</Nominal><ParentCode>R01040    </ParentCode><ISO_Num_Code></ISO_Num_Code><ISO_Char_Code></ISO_Char_Code></Item><Item ID="R01040E"><Name>\xc0\xed\xe3\xee\xeb\xfc\xf1\xea\xe0\xff \xea\xe2\xe0\xed\xe7\xe0</Name><EngName>Angolan Kwanza</EngName><Nominal>100</Nominal><ParentCode>R01040    </ParentCode><ISO_Num_Code>973</ISO_Num_Code><ISO_Char_Code>AOA</ISO_Char_Code></Item><Item ID="R01055"><Name>\xc0\xf0\xe3\xe5\xed\xf2\xe8\xed\xf1\xea\xee\xe5 \xef\xe5\xf1\xee</Name><EngName>Argentine Peso</EngName><Nominal>10</Nominal><ParentCode>R01055    </ParentCode><ISO_Num_Code>32</ISO_Num_Code><ISO_Char_Code>ARS</ISO_Char_Code></Item><Item ID="R01065"><Name>\xc0\xf4\xe3\xe0\xed\xf1\xea\xe8\xe9 \xe0\xf4\xe3\xe0\xed\xe8</Name><EngName>Afghanistan Afgani</EngName><Nominal>100</Nominal><ParentCode>R01065    </ParentCode><ISO_Num_Code>971</ISO_Num_Code><ISO_Char_Code>AFN</ISO_Char_Code></Item><Item ID="R01080"><Name>\xc1\xe0\xf5\xf0\xe5\xe9\xed\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Bahraini Dinar</EngName><Nominal>1</Nominal><ParentCode>R01080    </ParentCode><ISO_Num_Code>48</ISO_Num_Code><ISO_Char_Code>BHD</ISO_Char_Code></Item><Item ID="R01105"><Name>\xc1\xee\xeb\xe8\xe2\xe8\xe9\xf1\xea\xe8\xe9 \xe1\xee\xeb\xe8\xe2\xe8\xe0\xed\xee</Name><EngName>Bolivian Boliviano</EngName><Nominal>10</Nominal><ParentCode>R01105    </ParentCode><ISO_Num_Code>68</ISO_Num_Code><ISO_Char_Code>BOB</ISO_Char_Code></Item><Item ID="R01110"><Name>\xc1\xee\xf2\xf1\xe2\xe0\xed\xf1\xea\xe0\xff \xef\xf3\xeb\xe0</Name><EngName>Botswana Pula</EngName><Nominal>10</Nominal><ParentCode>R01110    </ParentCode><ISO_Num_Code>72</ISO_Num_Code><ISO_Char_Code>BWP</ISO_Char_Code></Item><Item ID="R01111"><Name>\xc1\xf0\xf3\xed\xe5\xe9\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Brunei Dollar</EngName><Nominal>1</Nominal><ParentCode>R01111    </ParentCode><ISO_Num_Code>96</ISO_Num_Code><ISO_Char_Code>BND</ISO_Char_Code></Item><Item ID="R01120"><Name>\xc1\xf3\xf0\xf3\xed\xe4\xe8\xe9\xf1\xea\xe8\xe9 \xf4\xf0\xe0\xed\xea</Name><EngName>Burundi Franc</EngName><Nominal>1000</Nominal><ParentCode>R01120    </ParentCode><ISO_Num_Code>108</ISO_Num_Code><ISO_Char_Code>BIF</ISO_Char_Code></Item><Item ID="R01140C"><Name>\xc2\xe5\xed\xe5\xf1\xf3\xfd\xeb\xfc\xf1\xea\xe8\xe9 \xe1\xee\xeb\xe8\xe2\xe0\xf0 \xf4\xf3\xfd\xf0\xf2\xe5</Name><EngName>Venezuela Bolivar Fuerte</EngName><Nominal>1</Nominal><ParentCode>R01140    </ParentCode><ISO_Num_Code>937</ISO_Num_Code><ISO_Char_Code>VEF</ISO_Char_Code></Item><Item ID="R01145"><Name>\xc2\xee\xed\xe0 \xca\xcd\xc4\xd0</Name><EngName>North Korean Won</EngName><Nominal>100</Nominal><ParentCode>R01145    </ParentCode><ISO_Num_Code>408</ISO_Num_Code><ISO_Char_Code>KPW</ISO_Char_Code></Item><Item ID="R01150"><Name>\xc2\xfc\xe5\xf2\xed\xe0\xec\xf1\xea\xe8\xe9 \xe4\xee\xed\xe3</Name><EngName>Vietnam Dong</EngName><Nominal>10000</Nominal><ParentCode>R01150    </ParentCode><ISO_Num_Code>704</ISO_Num_Code><ISO_Char_Code>VND</ISO_Char_Code></Item><Item ID="R01160"><Name>\xc3\xe0\xec\xe1\xe8\xe9\xf1\xea\xe8\xe9 \xe4\xe0\xeb\xe0\xf1\xe8</Name><EngName>Gambian Dalasi</EngName><Nominal>10</Nominal><ParentCode>R01160    </ParentCode><ISO_Num_Code>270</ISO_Num_Code><ISO_Char_Code>GMD</ISO_Char_Code></Item><Item ID="R01165D"><Name>\xc3\xe0\xed\xf1\xea\xe8\xe9 \xf1\xe5\xe4\xe8</Name><EngName>Ghana Cedi</EngName><Nominal>1</Nominal><ParentCode>R01165    </ParentCode><ISO_Num_Code>936</ISO_Num_Code><ISO_Char_Code>GHS</ISO_Char_Code></Item><Item ID="R01175"><Name>\xc3\xe2\xe8\xed\xe5\xe9\xf1\xea\xe8\xe9 \xf4\xf0\xe0\xed\xea</Name><EngName>Guinea Franc</EngName><Nominal>10000</Nominal><ParentCode>R01175    </ParentCode><ISO_Num_Code>324</ISO_Num_Code><ISO_Char_Code>GNF</ISO_Char_Code></Item><Item ID="R01210"><Name>\xc3\xf0\xf3\xe7\xe8\xed\xf1\xea\xe8\xe9 \xeb\xe0\xf0\xe8</Name><EngName>Georgia Lari</EngName><Nominal>1</Nominal><ParentCode>R01210    </ParentCode><ISO_Num_Code>981</ISO_Num_Code><ISO_Char_Code>GEL</ISO_Char_Code></Item><Item ID="R01230"><Name>\xc4\xe8\xf0\xf5\xe0\xec \xce\xc0\xdd</Name><EngName>UAE Dirham</EngName><Nominal>10</Nominal><ParentCode>R01230    </ParentCode><ISO_Num_Code>784</ISO_Num_Code><ISO_Char_Code>AED</ISO_Char_Code></Item><Item ID="R01233"><Name>\xc4\xee\xeb\xeb\xe0\xf0 \xc7\xe8\xec\xe1\xe0\xe1\xe2\xe5</Name><EngName>Zimbabwe Dollar</EngName><Nominal>100000</Nominal><ParentCode>R01233    </ParentCode><ISO_Num_Code>716</ISO_Num_Code><ISO_Char_Code>ZWD</ISO_Char_Code></Item><Item ID="R01233C"><Name>\xcd\xee\xe2\xfb\xe9 \xe4\xee\xeb\xeb\xe0\xf0 \xc7\xe8\xec\xe1\xe0\xe1\xe2\xe5</Name><EngName>New Zimbabwe dollar</EngName><Nominal>100</Nominal><ParentCode>R01233    </ParentCode><ISO_Num_Code>942</ISO_Num_Code><ISO_Char_Code>ZWN</ISO_Char_Code></Item><Item ID="R01240"><Name>\xc5\xe3\xe8\xef\xe5\xf2\xf1\xea\xe8\xe9 \xf4\xf3\xed\xf2</Name><EngName>Egyptian Pound</EngName><Nominal>10</Nominal><ParentCode>R01240    </ParentCode><ISO_Num_Code>818</ISO_Num_Code><ISO_Char_Code>EGP</ISO_Char_Code></Item><Item ID="R01245"><Name>\xc7\xe0\xe8\xf0 \xc4\xd0\xca</Name><EngName>Congo (Dem.Rep.) Zaire</EngName><Nominal>100000</Nominal><ParentCode>R01245    </ParentCode><ISO_Num_Code>180</ISO_Num_Code><ISO_Char_Code>ZRN</ISO_Char_Code></Item><Item ID="R01250"><Name>\xc7\xe0\xec\xe1\xe8\xe9\xf1\xea\xe0\xff \xea\xe2\xe0\xf7\xe0</Name><EngName>Zambian Kwacha</EngName><Nominal>10000</Nominal><ParentCode>R01250    </ParentCode><ISO_Num_Code>894</ISO_Num_Code><ISO_Char_Code>ZMK</ISO_Char_Code></Item><Item ID="R01265"><Name>\xc8\xe7\xf0\xe0\xe8\xeb\xfc\xf1\xea\xe8\xe9 \xed\xee\xe2\xfb\xe9 \xf8\xe5\xea\xe5\xeb\xfc</Name><EngName>Izraeli Shekel</EngName><Nominal>10</Nominal><ParentCode>R01265    </ParentCode><ISO_Num_Code>376</ISO_Num_Code><ISO_Char_Code>ILS</ISO_Char_Code></Item><Item ID="R01265C"><Name>\xcd\xee\xe2\xfb\xe9 \xe8\xe7\xf0\xe0\xe8\xeb\xfc\xf1\xea\xe8\xe9 \xf8\xe5\xea\xe5\xeb\xfc</Name><EngName>Izraeli Shekel</EngName><Nominal>10</Nominal><ParentCode>R01265    </ParentCode><ISO_Num_Code>376</ISO_Num_Code><ISO_Char_Code>ILS</ISO_Char_Code></Item><Item ID="R01280"><Name>\xc8\xed\xe4\xee\xed\xe5\xe7\xe8\xe9\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Indonesian Rupiah</EngName><Nominal>10000</Nominal><ParentCode>R01280    </ParentCode><ISO_Num_Code>360</ISO_Num_Code><ISO_Char_Code>IDR</ISO_Char_Code></Item><Item ID="R01285"><Name>\xc8\xee\xf0\xe4\xe0\xed\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Jordanian Dinar</EngName><Nominal>1</Nominal><ParentCode>R01285    </ParentCode><ISO_Num_Code>400</ISO_Num_Code><ISO_Char_Code>JOD</ISO_Char_Code></Item><Item ID="R01290"><Name>\xc8\xf0\xe0\xea\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Iraqi Dinar</EngName><Nominal>1000</Nominal><ParentCode>R01290    </ParentCode><ISO_Num_Code>368</ISO_Num_Code><ISO_Char_Code>IQD</ISO_Char_Code></Item><Item ID="R01300"><Name>\xc8\xf0\xe0\xed\xf1\xea\xe8\xe9 \xf0\xe8\xe0\xeb</Name><EngName>Iranian Rial</EngName><Nominal>10000</Nominal><ParentCode>R01300    </ParentCode><ISO_Num_Code>364</ISO_Num_Code><ISO_Char_Code>IRR</ISO_Char_Code></Item><Item ID="R01330"><Name>\xc9\xe5\xec\xe5\xed\xf1\xea\xe8\xe9 \xf0\xe8\xe0\xeb</Name><EngName>Yemeni Rial</EngName><Nominal>100</Nominal><ParentCode>R01330    </ParentCode><ISO_Num_Code>886</ISO_Num_Code><ISO_Char_Code>YER</ISO_Char_Code></Item><Item ID="R01355"><Name>\xca\xe0\xf2\xe0\xf0\xf1\xea\xe8\xe9 \xf0\xe8\xe0\xeb</Name><EngName>Qatari Riyal</EngName><Nominal>10</Nominal><ParentCode>R01355    </ParentCode><ISO_Num_Code>634</ISO_Num_Code><ISO_Char_Code>QAR</ISO_Char_Code></Item><Item ID="R01360"><Name>\xca\xe5\xed\xe8\xe9\xf1\xea\xe8\xe9 \xf8\xe8\xeb\xeb\xe8\xed\xe3</Name><EngName>Kenyan Shilling</EngName><Nominal>100</Nominal><ParentCode>R01360    </ParentCode><ISO_Num_Code>404</ISO_Num_Code><ISO_Char_Code>KES</ISO_Char_Code></Item><Item ID="R01365"><Name>\xca\xe8\xef\xf0\xf1\xea\xe8\xe9 \xf4\xf3\xed\xf2</Name><EngName>Cypriot Pound</EngName><Nominal>1</Nominal><ParentCode>R01365    </ParentCode><ISO_Num_Code>196</ISO_Num_Code><ISO_Char_Code>CYP</ISO_Char_Code></Item><Item ID="R01380"><Name>\xca\xee\xeb\xf3\xec\xe1\xe8\xe9\xf1\xea\xe8\xe9 \xef\xe5\xf1\xee</Name><EngName>Colombian Peso</EngName><Nominal>1000</Nominal><ParentCode>R01380    </ParentCode><ISO_Num_Code>170</ISO_Num_Code><ISO_Char_Code>COP</ISO_Char_Code></Item><Item ID="R01383"><Name>\xca\xee\xed\xe3\xee\xeb\xe5\xe7\xf1\xea\xe8\xe9 \xf4\xf0\xe0\xed\xea</Name><EngName>Congolese franc</EngName><Nominal>1000</Nominal><ParentCode>R01383    </ParentCode><ISO_Num_Code>976</ISO_Num_Code><ISO_Char_Code>CDF</ISO_Char_Code></Item><Item ID="R01385"><Name>\xca\xee\xf1\xf2\xe0\xf0\xe8\xea\xe0\xed\xf1\xea\xe8\xe9 \xea\xee\xeb\xee\xed</Name><EngName>Costa Rican colon</EngName><Nominal>1000</Nominal><ParentCode>R01385    </ParentCode><ISO_Num_Code>188</ISO_Num_Code><ISO_Char_Code>CRC</ISO_Char_Code></Item><Item ID="R01395"><Name>\xca\xf3\xe1\xe8\xed\xf1\xea\xee\xe5 \xef\xe5\xf1\xee</Name><EngName>Cuban peso</EngName><Nominal>1</Nominal><ParentCode>R01395    </ParentCode><ISO_Num_Code>192</ISO_Num_Code><ISO_Char_Code>CUP</ISO_Char_Code></Item><Item ID="R01400"><Name>\xcb\xe0\xee\xf1\xf1\xea\xe8\xe9 \xea\xe8\xef</Name><EngName>Laos new Kip</EngName><Nominal>10000</Nominal><ParentCode>R01400    </ParentCode><ISO_Num_Code>418</ISO_Num_Code><ISO_Char_Code>LAK</ISO_Char_Code></Item><Item ID="R01410"><Name>\xcb\xe5\xee\xed\xe5 \xd1\xfc\xe5\xf0\xf0\xe0-\xcb\xe5\xee\xed\xe5</Name><EngName>Sierra Leone Leone</EngName><Nominal>10000</Nominal><ParentCode>R01410    </ParentCode><ISO_Num_Code>694</ISO_Num_Code><ISO_Char_Code>SLL</ISO_Char_Code></Item><Item ID="R01425"><Name>\xcb\xe8\xe2\xe8\xe9\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Libyan Dinar</EngName><Nominal>1</Nominal><ParentCode>R01425    </ParentCode><ISO_Num_Code>434</ISO_Num_Code><ISO_Char_Code>LYD</ISO_Char_Code></Item><Item ID="R01430"><Name>\xcb\xe8\xeb\xe0\xed\xe3\xe5\xed\xe8 \xd1\xe2\xe0\xe7\xe8\xeb\xe5\xed\xe4\xe0</Name><EngName>Swaziland Lilangeni</EngName><Nominal>10</Nominal><ParentCode>R01430    </ParentCode><ISO_Num_Code>748</ISO_Num_Code><ISO_Char_Code>SZL</ISO_Char_Code></Item><Item ID="R01445"><Name>\xcc\xe0\xe2\xf0\xe8\xea\xe8\xe9\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Mauritius Rupee</EngName><Nominal>100</Nominal><ParentCode>R01445    </ParentCode><ISO_Num_Code>480</ISO_Num_Code><ISO_Char_Code>MUR</ISO_Char_Code></Item><Item ID="R01450"><Name>\xcc\xe0\xe2\xf0\xe8\xf2\xe0\xed\xf1\xea\xe0\xff \xf3\xe3\xe8\xff</Name><EngName>Mauritania Ouguiya</EngName><Nominal>100</Nominal><ParentCode>R01450    </ParentCode><ISO_Num_Code>478</ISO_Num_Code><ISO_Char_Code>MRO</ISO_Char_Code></Item><Item ID="R01460"><Name>\xcc\xe0\xea\xe5\xe4\xee\xed\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Macedonia Denar</EngName><Nominal>100</Nominal><ParentCode>R01460    </ParentCode><ISO_Num_Code>807</ISO_Num_Code><ISO_Char_Code>MKD</ISO_Char_Code></Item><Item ID="R01465"><Name>\xcc\xe0\xeb\xe0\xe2\xe8\xe9\xf1\xea\xe0\xff \xea\xe2\xe0\xf7\xe0</Name><EngName>Malawi Kwacha</EngName><Nominal>100</Nominal><ParentCode>R01465    </ParentCode><ISO_Num_Code>454</ISO_Num_Code><ISO_Char_Code>MWK</ISO_Char_Code></Item><Item ID="R01470C"><Name>\xcc\xe0\xeb\xe0\xe3\xe0\xf1\xe8\xe9\xf1\xea\xe8\xe9 \xe0\xf0\xe8\xe0\xf0\xe8</Name><EngName>Malagasy ariary</EngName><Nominal>1000</Nominal><ParentCode>R01470    </ParentCode><ISO_Num_Code>969</ISO_Num_Code><ISO_Char_Code>MGA</ISO_Char_Code></Item><Item ID="R01475"><Name>\xcc\xe0\xeb\xe0\xe9\xe7\xe8\xe9\xf1\xea\xe8\xe9 \xf0\xe8\xed\xe3\xe3\xe8\xf2</Name><EngName>Malaysian Ringgit</EngName><Nominal>10</Nominal><ParentCode>R01475    </ParentCode><ISO_Num_Code>458</ISO_Num_Code><ISO_Char_Code>MYR</ISO_Char_Code></Item><Item ID="R01480"><Name>\xcc\xe0\xeb\xfc\xf2\xe8\xe9\xf1\xea\xe0\xff \xeb\xe8\xf0\xe0</Name><EngName>Maltese Lira</EngName><Nominal>1</Nominal><ParentCode>R01480    </ParentCode><ISO_Num_Code>470</ISO_Num_Code><ISO_Char_Code>MTL</ISO_Char_Code></Item><Item ID="R01485"><Name>\xcc\xe0\xf0\xee\xea\xea\xe0\xed\xf1\xea\xe8\xe9 \xe4\xe8\xf0\xf5\xe0\xec</Name><EngName>Moroccan Dirham</EngName><Nominal>10</Nominal><ParentCode>R01485    </ParentCode><ISO_Num_Code>504</ISO_Num_Code><ISO_Char_Code>MAD</ISO_Char_Code></Item><Item ID="R01495"><Name>\xcc\xe5\xea\xf1\xe8\xea\xe0\xed\xf1\xea\xe8\xe9 \xef\xe5\xf1\xee</Name><EngName>Mexican Peso</EngName><Nominal>10</Nominal><ParentCode>R01495    </ParentCode><ISO_Num_Code>484</ISO_Num_Code><ISO_Char_Code>MXN</ISO_Char_Code></Item><Item ID="R01498C"><Name>\xcc\xee\xe7\xe0\xec\xe1\xe8\xea\xf1\xea\xe8\xe9 \xec\xe5\xf2\xe8\xea\xe0\xeb</Name><EngName>Mozambique Metical</EngName><Nominal>100</Nominal><ParentCode>R01498    </ParentCode><ISO_Num_Code>943</ISO_Num_Code><ISO_Char_Code>MZN</ISO_Char_Code></Item><Item ID="R01503"><Name>\xcc\xee\xed\xe3\xee\xeb\xfc\xf1\xea\xe8\xe9 \xf2\xf3\xe3\xf0\xe8\xea</Name><EngName>Mongolia Tugrik</EngName><Nominal>1000</Nominal><ParentCode>R01503    </ParentCode><ISO_Num_Code>496</ISO_Num_Code><ISO_Char_Code>MNT</ISO_Char_Code></Item><Item ID="R01515"><Name>\xcd\xe5\xef\xe0\xeb\xfc\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Nepal Ruppe</EngName><Nominal>100</Nominal><ParentCode>R01515    </ParentCode><ISO_Num_Code>524</ISO_Num_Code><ISO_Char_Code>NPR</ISO_Char_Code></Item><Item ID="R01520"><Name>\xcd\xe8\xe3\xe5\xf0\xe8\xe9\xf1\xea\xe8\xe9 \xed\xe0\xe9\xf0</Name><EngName>Nigeria Naira</EngName><Nominal>100</Nominal><ParentCode>R01520    </ParentCode><ISO_Num_Code>566</ISO_Num_Code><ISO_Char_Code>NGN</ISO_Char_Code></Item><Item ID="R01525"><Name>\xcd\xe8\xea\xe0\xf0\xe0\xe3\xf3\xe0\xed\xf1\xea\xe0\xff \xe7\xee\xeb\xee\xf2\xe0\xff \xea\xee\xf0\xe4\xee\xe1\xe0</Name><EngName>Nicaragua Cordoba</EngName><Nominal>10</Nominal><ParentCode>R01525    </ParentCode><ISO_Num_Code>558</ISO_Num_Code><ISO_Char_Code>NIO</ISO_Char_Code></Item><Item ID="R01530"><Name>\xcd\xee\xe2\xee\xe7\xe5\xeb\xe0\xed\xe4\xf1\xea\xe8\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>New Zealand Dollar</EngName><Nominal>1</Nominal><ParentCode>R01530    </ParentCode><ISO_Num_Code>554</ISO_Num_Code><ISO_Char_Code>NZD</ISO_Char_Code></Item><Item ID="R01540"><Name>\xce\xec\xe0\xed\xf1\xea\xe8\xe9 \xf0\xe8\xe0\xeb</Name><EngName>Omani Rial</EngName><Nominal>1</Nominal><ParentCode>R01540    </ParentCode><ISO_Num_Code>512</ISO_Num_Code><ISO_Char_Code>OMR</ISO_Char_Code></Item><Item ID="R01545"><Name>\xcf\xe0\xea\xe8\xf1\xf2\xe0\xed\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Pakistani Rupee</EngName><Nominal>100</Nominal><ParentCode>R01545    </ParentCode><ISO_Num_Code>586</ISO_Num_Code><ISO_Char_Code>PKR</ISO_Char_Code></Item><Item ID="R01555"><Name>\xcf\xe0\xf0\xe0\xe3\xe2\xe0\xe9\xf1\xea\xe0\xff \xe3\xf3\xe0\xf0\xe0\xed\xe8</Name><EngName>Paraguay Guarani</EngName><Nominal>10000</Nominal><ParentCode>R01555    </ParentCode><ISO_Num_Code>600</ISO_Num_Code><ISO_Char_Code>PYG</ISO_Char_Code></Item><Item ID="R01560"><Name>\xcf\xe5\xf0\xf3\xe0\xed\xf1\xea\xe8\xe9 \xed\xee\xe2\xfb\xe9 \xf1\xee\xeb\xfc</Name><EngName>Peruviann new Sol</EngName><Nominal>1</Nominal><ParentCode>R01560    </ParentCode><ISO_Num_Code>604</ISO_Num_Code><ISO_Char_Code>PEN</ISO_Char_Code></Item><Item ID="R01575"><Name>\xd0\xe8\xe5\xeb\xfc \xca\xe0\xec\xe1\xee\xe4\xe6\xe8</Name><EngName>Cambodia Riel</EngName><Nominal>10000</Nominal><ParentCode>R01575    </ParentCode><ISO_Num_Code>116</ISO_Num_Code><ISO_Char_Code>KHR</ISO_Char_Code></Item><Item ID="R01580"><Name>\xd0\xe8\xff\xeb \xd1\xe0\xf3\xe4\xee\xe2\xf1\xea\xee\xe9 \xc0\xf0\xe0\xe2\xe8\xe8</Name><EngName>Saudi Riyal</EngName><Nominal>10</Nominal><ParentCode>R01580    </ParentCode><ISO_Num_Code>682</ISO_Num_Code><ISO_Char_Code>SAR</ISO_Char_Code></Item><Item ID="R01580C"><Name>\xd1\xe0\xf3\xe4\xee\xe2\xf1\xea\xe8\xe9 \xf0\xe8\xff\xeb</Name><EngName>Saudi Riyal</EngName><Nominal>10</Nominal><ParentCode>R01580    </ParentCode><ISO_Num_Code>682</ISO_Num_Code><ISO_Char_Code>SAR</ISO_Char_Code></Item><Item ID="R01595"><Name>\xd1\xe5\xe9\xf8\xe5\xeb\xfc\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Seychelles Rupee</EngName><Nominal>10</Nominal><ParentCode>R01595    </ParentCode><ISO_Num_Code>690</ISO_Num_Code><ISO_Char_Code>SCR</ISO_Char_Code></Item><Item ID="R01630"><Name>\xd1\xe8\xf0\xe8\xe9\xf1\xea\xe8\xe9 \xf4\xf3\xed\xf2</Name><EngName>Syrian Pound</EngName><Nominal>100</Nominal><ParentCode>R01630    </ParentCode><ISO_Num_Code>760</ISO_Num_Code><ISO_Char_Code>SYP</ISO_Char_Code></Item><Item ID="R01635"><Name>\xd1\xeb\xee\xe2\xe0\xf6\xea\xe0\xff \xea\xf0\xee\xed\xe0</Name><EngName>Slovakia Koruna</EngName><Nominal>10</Nominal><ParentCode>R01635    </ParentCode><ISO_Num_Code>703</ISO_Num_Code><ISO_Char_Code>SKK</ISO_Char_Code></Item><Item ID="R01640"><Name>\xd1\xeb\xee\xe2\xe5\xed\xf1\xea\xe8\xe9 \xf2\xee\xeb\xe0\xf0</Name><EngName>Slovenia Tolar</EngName><Nominal>100</Nominal><ParentCode>R01640    </ParentCode><ISO_Num_Code>705</ISO_Num_Code><ISO_Char_Code>SIT</ISO_Char_Code></Item><Item ID="R01650"><Name>\xd1\xee\xec\xe0\xeb\xe8\xe9\xf1\xea\xe8\xe9 \xf8\xe8\xeb\xeb\xe8\xed\xe3</Name><EngName>Somali Schilling</EngName><Nominal>1000</Nominal><ParentCode>R01650    </ParentCode><ISO_Num_Code>706</ISO_Num_Code><ISO_Char_Code>SOS</ISO_Char_Code></Item><Item ID="R01660A"><Name>\xd1\xf3\xe4\xe0\xed\xf1\xea\xe8\xe9 \xf4\xf3\xed\xf2</Name><EngName>Sudanese Pound</EngName><Nominal>1</Nominal><ParentCode>R01660    </ParentCode><ISO_Num_Code>938</ISO_Num_Code><ISO_Char_Code>SDG</ISO_Char_Code></Item><Item ID="R01675"><Name>\xd2\xe0\xe8\xeb\xe0\xed\xe4\xf1\xea\xe8\xe9 \xe1\xe0\xf2</Name><EngName>Thai Baht</EngName><Nominal>100</Nominal><ParentCode>R01675    </ParentCode><ISO_Num_Code>764</ISO_Num_Code><ISO_Char_Code>THB</ISO_Char_Code></Item><Item ID="R01680"><Name>\xd2\xe0\xe9\xe2\xe0\xed\xfc\xf1\xea\xe8\xe9 \xed\xee\xe2\xfb\xe9 \xe4\xee\xeb\xeb\xe0\xf0</Name><EngName>Taiwan Dollar</EngName><Nominal>100</Nominal><ParentCode>R01680    </ParentCode><ISO_Num_Code>901</ISO_Num_Code><ISO_Char_Code>TWD</ISO_Char_Code></Item><Item ID="R01685"><Name>\xd2\xe0\xea \xc1\xe0\xed\xe3\xeb\xe0\xe4\xe5\xf8</Name><EngName>Bangladesh Taka</EngName><Nominal>100</Nominal><ParentCode>R01685    </ParentCode><ISO_Num_Code>50</ISO_Num_Code><ISO_Char_Code>BDT</ISO_Char_Code></Item><Item ID="R01690"><Name>\xd2\xe0\xed\xe7\xe0\xed\xe8\xe9\xf1\xea\xe8\xe9 \xf8\xe8\xeb\xeb\xe8\xed\xe3</Name><EngName>Tanzanian Shilling</EngName><Nominal>1000</Nominal><ParentCode>R01690    </ParentCode><ISO_Num_Code>834</ISO_Num_Code><ISO_Char_Code>TZS</ISO_Char_Code></Item><Item ID="R01695"><Name>\xd2\xf3\xed\xe8\xf1\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Tunisian Dinar</EngName><Nominal>1</Nominal><ParentCode>R01695    </ParentCode><ISO_Num_Code>788</ISO_Num_Code><ISO_Char_Code>TND</ISO_Char_Code></Item><Item ID="R01714"><Name>\xd3\xe3\xe0\xed\xe4\xe8\xe9\xf1\xea\xe8\xe9 \xf8\xe8\xeb\xeb\xe8\xed\xe3</Name><EngName>Ugandan Shilling</EngName><Nominal>1000</Nominal><ParentCode>R01714    </ParentCode><ISO_Num_Code>800</ISO_Num_Code><ISO_Char_Code>UGX</ISO_Char_Code></Item><Item ID="R01725"><Name>\xd3\xf0\xf3\xe3\xe2\xe0\xe9\xf1\xea\xee\xe5 \xef\xe5\xf1\xee</Name><EngName>Uruguay Peso</EngName><Nominal>10</Nominal><ParentCode>R01725    </ParentCode><ISO_Num_Code>858</ISO_Num_Code><ISO_Char_Code>UYU</ISO_Char_Code></Item><Item ID="R01743"><Name>\xd4\xe8\xeb\xe8\xef\xef\xe8\xed\xf1\xea\xee\xe5 \xef\xe5\xf1\xee</Name><EngName>Philippines Peso</EngName><Nominal>100</Nominal><ParentCode>R01743    </ParentCode><ISO_Num_Code>608</ISO_Num_Code><ISO_Char_Code>PHP</ISO_Char_Code></Item><Item ID="R01746"><Name>\xd4\xf0\xe0\xed\xea \xc4\xe6\xe8\xe1\xf3\xf2\xe8</Name><EngName>Djobouti Franc</EngName><Nominal>100</Nominal><ParentCode>R01746    </ParentCode><ISO_Num_Code>262</ISO_Num_Code><ISO_Char_Code>DJF</ISO_Char_Code></Item><Item ID="R01748"><Name>\xd4\xf0\xe0\xed\xea \xca\xd4\xc0 \xc2\xc5\xc0\xd1</Name><EngName>CFA frank BEAC</EngName><Nominal>1000</Nominal><ParentCode>R01748    </ParentCode><ISO_Num_Code>950</ISO_Num_Code><ISO_Char_Code>XAF</ISO_Char_Code></Item><Item ID="R01749"><Name>\xd4\xf0\xe0\xed\xea \xca\xd4\xc0 \xc2\xd1\xc5\xc0\xce</Name><EngName>CFA frank BCEAO</EngName><Nominal>1000</Nominal><ParentCode>R01749    </ParentCode><ISO_Num_Code>952</ISO_Num_Code><ISO_Char_Code>XOF</ISO_Char_Code></Item><Item ID="R01755"><Name>\xd5\xee\xf0\xe2\xe0\xf2\xf1\xea\xe0\xff \xea\xf3\xed\xe0</Name><EngName>Croatian Kuna</EngName><Nominal>10</Nominal><ParentCode>R01755    </ParentCode><ISO_Num_Code>191</ISO_Num_Code><ISO_Char_Code>HRK</ISO_Char_Code></Item><Item ID="R01765"><Name>\xd7\xe8\xeb\xe8\xe9\xf1\xea\xee\xe5 \xef\xe5\xf1\xee</Name><EngName>Chilean Peso</EngName><Nominal>1000</Nominal><ParentCode>R01765    </ParentCode><ISO_Num_Code>152</ISO_Num_Code><ISO_Char_Code>CLP</ISO_Char_Code></Item><Item ID="R01780"><Name>\xd8\xf0\xe8-\xcb\xe0\xed\xea\xe8\xe9\xf1\xea\xe0\xff \xf0\xf3\xef\xe8\xff</Name><EngName>Sri Lankan Rupee</EngName><Nominal>100</Nominal><ParentCode>R01780    </ParentCode><ISO_Num_Code>144</ISO_Num_Code><ISO_Char_Code>LKR</ISO_Char_Code></Item><Item ID="R01785"><Name>\xdd\xea\xe2\xe0\xe4\xee\xf0\xf1\xea\xe8\xe9 \xf1\xf3\xea\xf0\xe5</Name><EngName>Ecuadoran Sucre</EngName><Nominal>1000</Nominal><ParentCode>R01785    </ParentCode><ISO_Num_Code>218</ISO_Num_Code><ISO_Char_Code>ECS</ISO_Char_Code></Item><Item ID="R01800"><Name>\xdd\xf4\xe8\xee\xef\xf1\xea\xe8\xe9 \xe1\xfb\xf0</Name><EngName>Ethiopian Birr</EngName><Nominal>10</Nominal><ParentCode>R01800    </ParentCode><ISO_Num_Code>230</ISO_Num_Code><ISO_Char_Code>ETB</ISO_Char_Code></Item><Item ID="R01804"><Name>\xd1\xe5\xf0\xe1\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Serbian Dinar</EngName><Nominal>100</Nominal><ParentCode>R01804    </ParentCode><ISO_Num_Code></ISO_Num_Code><ISO_Char_Code></ISO_Char_Code></Item><Item ID="R01805F"><Name>\xd1\xe5\xf0\xe1\xf1\xea\xe8\xe9 \xe4\xe8\xed\xe0\xf0</Name><EngName>Serbian Dinar</EngName><Nominal>100</Nominal><ParentCode>R01804    </ParentCode><ISO_Num_Code>941</ISO_Num_Code><ISO_Char_Code>RSD</ISO_Char_Code></Item><Item ID="R02004"><Name>\xc4\xee\xeb\xeb\xe0\xf0 \xcd\xe0\xec\xe8\xe1\xe8\xe8</Name><EngName>Namibian dollar</EngName><Nominal>10</Nominal><ParentCode>R02004    </ParentCode><ISO_Num_Code>516</ISO_Num_Code><ISO_Char_Code>NAD</ISO_Char_Code></Item></Valuta>'
