+++ pycbrf/rates.py
@@ -1,49 +1,15 @@
-import datetime as dt
+from datetime import date, datetime
 from decimal import Decimal
 from logging import getLogger
 from typing import Dict, NamedTuple, Optional, Tuple, Union
 from xml.etree import ElementTree
 
+from .constants import URL_BASE
+from .currencies import Currency, Currencies
 from .exceptions import CurrencyNotFound, ExchangeRateNotFound, WrongArguments
-from .utils import SingletonMeta, FormatMixin, WithRequests
+from .utils import FormatMixin, WithRequests
 
 LOG = getLogger(__name__)
-URL_BASE = 'http://www.cbr.ru/scripts/'
-
-
-class Currency(NamedTuple):
-    """Represents a foreign currency.
-
-    :param id: Internal code of the Bank of Russia.
-    :param name_eng: Currency name in English.
-    :param name_ru: Currency name in Russian.
-    :param par: Nominal exchange rate.
-    :param num: ISO 4217 currency numeric code.
-    :param code: ISO 4217 currency alphabetic code.
-
-    :Example:
-
-    Currency(
-        id='R01010',
-        name_ru='Австралийский доллар',
-        name_eng='Australian Dollar',
-        num='036',
-        code='AUD',
-        par=Decimal('1'))
-    """
-
-    id: str
-    name_eng: str
-    name_ru: str
-    num: str
-    code: str
-    par: Decimal
-
-    def __hash__(self):
-        return hash((self.id, self.num, self.code))
-
-    def __eq__(self, cls):
-        return isinstance(cls, type(self)) and (cls.id, cls.num, cls.code) == (self.id, self.num, self.code)
 
 
 class ExchangeRate(NamedTuple):
@@ -56,7 +22,7 @@ class ExchangeRate(NamedTuple):
     :param rate: Reduced exchange rate rate = value / par.
     """
 
-    date: dt.datetime
+    on_date: datetime
     currency: Currency
     name: str  # exists for backward compatibility.
     value: Decimal
@@ -88,17 +54,11 @@ class ExchangeRates(WithRequests, FormatMixin):
     """Gets and stores exchange rates for different currencies for a selected date
 
     :param date_requested: The date on which the rates are requested.
-    :type date_requested: date
     :param date_received: The date of the rates that was returned from the Central Bank of Russia.
-    :type date_received: date
     :param dates_match: Returns whether the requested rate date and response rate date are the same.
-    :type dates_match: bool
     :param rates: Dictionary of ExchangeRate parsed from the server.
-    :type rates: Dict[Currency, ExchangeRate]
     :param length: Number of exchange rates.
-    :type length: int
     :param currencies_lib: Library of currencies.
-    :type currencies_lib: Currencies
 
     :Example:
 
@@ -117,12 +77,10 @@ class ExchangeRates(WithRequests, FormatMixin):
         The rates of holiday is set as equal on the last working day before the holidays.
     """
 
-    def __init__(self, on_date: Union[str, dt.date, dt.datetime, None] = None, locale_en: bool = False):
+    def __init__(self, on_date: Union[str, date, datetime, None] = None, locale_en: bool = False):
         """
         :param on_date: Date to get exchange rates for.
-        :type on_date: str, date, datetime, optional
         :param locale_en: If not set ExchangeRate.name will be provided in Russian, otherwise in English.
-        :type locale_en: bool
         """
         self.currencies_lib = Currencies()
         self.length = 0
@@ -130,15 +88,15 @@ class ExchangeRates(WithRequests, FormatMixin):
         if on_date:
             on_date = self._datetime_from_string(on_date)
         else:
-            today = dt.date.today()
-            on_date = dt.datetime(today.year, today.month, today.day)  # For backward compatibility, the time is 00:00
+            today = date.today()
+            on_date = datetime(today.year, today.month, today.day)  # For backward compatibility, the time is 00:00
 
         self.date_requested = on_date
 
         raw_data = self._get_data(on_date, locale_en)
         date_received, rates = self._parse(raw_data, locale_en)
 
-        self.date_received: dt.datetime = date_received
+        self.date_received: datetime = date_received
         self.rates: Dict[Currency, ExchangeRate] = rates
         self.dates_match: bool = (self.date_requested == self.date_received)
 
@@ -146,9 +104,7 @@ class ExchangeRates(WithRequests, FormatMixin):
         """Implement dictionary lookup
 
         :param item: Bank of Russia code, numeric or alphabetic currency code according to ISO, Currency instance
-        :type item: str, int, Currency
         :return: The ExchangeRate instance for the requested currency
-        :rtype: ExchangeRate, optional
         """
         try:
             key: Optional[Currency] = self.currencies_lib[item]
@@ -164,7 +120,7 @@ class ExchangeRates(WithRequests, FormatMixin):
         return len(self.rates)
 
     @classmethod
-    def _get_data(cls, on_date: dt.datetime, locale_en: bool) -> bytes:
+    def _get_data(cls, on_date: datetime, locale_en: bool) -> bytes:
         """Prepares parameters for the link and returns raw XML"""
         url = f"{URL_BASE}XML_daily{'_eng' if locale_en else ''}.asp"
         params = {
@@ -176,13 +132,13 @@ class ExchangeRates(WithRequests, FormatMixin):
 
         return raw_data
 
-    def _parse(self, data: bytes, locale_en: bool) -> Tuple[dt.datetime, Dict[Currency, ExchangeRate]]:
+    def _parse(self, data: bytes, locale_en: bool) -> Tuple[datetime, Dict[Currency, ExchangeRate]]:
         """Parse raw XML strings to the dict of BetaExchangeRates"""
         LOG.debug('Parsing data ...')
 
         xml = ElementTree.fromstring(data)
 
-        date_received = dt.datetime.strptime(xml.attrib['Date'], '%d.%m.%Y')
+        date_received = datetime.strptime(xml.attrib['Date'], '%d.%m.%Y')
 
         result = {}
 
@@ -211,7 +167,7 @@ class ExchangeRates(WithRequests, FormatMixin):
             value = Decimal(params['Value'].replace(',', '.'))
 
             result[currency] = ExchangeRate(
-                date=date_received,
+                on_date=date_received,
                 currency=currency,
                 name=name,
                 value=value,
@@ -227,130 +183,15 @@ class ExchangeRates(WithRequests, FormatMixin):
         return f"ExchangeRates of {len(self.rates)} currencies from {self.date_requested}"
 
 
-class Currencies(WithRequests, FormatMixin, metaclass=SingletonMeta):
-    """Singleton class represents library of Currency
-
-    :param update_date: Date of loading the latest information from www.cbr.ru
-    :type update_date: dt.datetime
-    :param currencies: Dict of Currency
-    :type currencies: Dict[str, Currency]
-
-    .. note:: currencies represents {Union[Currency.id, Currency.num, Currency.code]: Currency} dictobject.
-    """
-
-    def __init__(self):
-        self.update_date = None
-        self.currencies = None
-        self.update()
-
-    def __getitem__(self, value: Union[int, str]) -> Optional[Currency]:
-        """Returns Currency by dictionary lookup, converting the argument to ISO format."""
-        if not value:
-            raise CurrencyNotFound(f'Currency "{value}" not found.')
-
-        item = self._format_num_code(value)
-        item = item.lower()
-
-        try:
-            currency = self.currencies[item]
-        except KeyError:
-            raise CurrencyNotFound()
-
-        return currency
-
-    def update(self):
-        """Get and parse actual data from the www.cbr.ru."""
-        raw_data = self._get_data()
-        self.currencies = self._parse(raw_data)
-        self.update_date = dt.datetime.now()
-
-    def add(self, currency: Currency):
-        """Add a new currency to the library or update an existing currency"""
-        if isinstance(currency, Currency):
-            self.currencies.update(self._make_currency_pack(currency))
-
-    @classmethod
-    def _get_data(cls) -> Tuple[bytes, bytes]:
-        """Get XML byte string from www.cbr.ru for dayly and monthly update currencies."""
-        url = f"{URL_BASE}XML_valFull.asp"
-
-        LOG.debug(f'Getting update currencies from {url} ...')
-
-        daily_update_response = cls._get_response(url)
-        daily_update_data = daily_update_response.content
-
-        monhtly_update_response = cls._get_response(url, params={'d': 1})
-        monthly_update_data = monhtly_update_response.content
-
-        return daily_update_data, monthly_update_data
-
-    @staticmethod
-    def _make_currency_pack(currency: Currency) -> Dict[str, Currency]:
-        """Creates a dict of three elements from the currency with the keys: id, code and num if they exists"""
-        pack = {}
-
-        pack[currency.id.lower()] = currency
-        # Data from the Bank of Russia contains replaced currencies that do not have ISO attributes.
-        # So additional If-statements were added to exclude None from the 'codes'.
-        if currency.code:
-            pack[currency.code.lower()] = currency
-        if currency.num:
-            pack[currency.num] = currency
-
-        return pack
-
-    def _parse(self, data: Tuple[bytes, bytes]) -> Dict[str, Currency]:
-        """Parse XML bytes strings from www.cbr.ru to dict of Currencies."""
-        currencies = {}
-        counter = 0
-
-        LOG.debug('Parsing data ...')
-
-        for sub_data in data:
-            root = ElementTree.fromstring(sub_data)
-
-            for child in root:
-                props = {}
-                for prop in child:
-                    props[prop.tag] = prop.text
-
-                currency = Currency(
-                    id=child.attrib['ID'],
-                    name_eng=props['EngName'],
-                    name_ru=props['Name'],
-                    code=props['ISO_Char_Code'],
-                    # ISO numeric code like '036' is loaded like '36', so it needs format to ISO 4217,
-                    # also data from the Bank of Russia contains replaced currencies that do not have ISO attributes.
-                    # additional If-statement was added to exclude format None
-                    num=self._format_num_code(_) if (_ := props['ISO_Num_Code']) else None,
-                    par=Decimal(props['Nominal']),
-                )
-
-                counter += 1
-                currencies.update(self._make_currency_pack(currency))
-
-        LOG.debug(f"Parsed: {counter} currencies")
-        return currencies
-
-    def __str__(self):
-        return f"Currencies. Update {self.update_date}"
-
-
 class ExchangeRateDynamics(WithRequests, FormatMixin):
     """Receives and stores exchange rates for one currency over a period of time
 
     :param date_from: Start date of the period for the exchange rate dynamics.
-    :type date_from: dt.datetime
     :param date_to: End date of the period for the exchange rate dynamics.
-    :type date_to: dt.datetime
     :param currency: The currency for which the exchange rates were requested.
-    :type currency: Currency
     :param rates: Dictionary of ExchangeRate parsed from the server.
-    :type rates: Dict[datetime, ExchangeRate])
     :param length: Number of exchange rates.
-    :type length: int
     :param currencies_lib: Library of currencies.
-    :type currencies_lib: Currencies
 
     :Example:
 
@@ -373,21 +214,18 @@ class ExchangeRateDynamics(WithRequests, FormatMixin):
     """
 
     def __init__(self,
-                 date_from: Union[str, dt.date, dt.datetime, None] = None,
-                 date_to: Union[str, dt.date, dt.datetime, None] = None,
+                 date_from: Union[str, date, datetime, None] = None,
+                 date_to: Union[str, date, datetime, None] = None,
                  *,
                  currency: Union[str, int, Currency],
                  ):
         """
         :param date_from: Date of the exchange rate or start date of the period for the exchange rate dynamics.
             Python date and datetime objects and ISO date '%Y-%m-%d' string are supported.
-        :type date_from: date, str, optional
         :param date_to: End date of the period for the exchange rate dynamics.
             Python date and datetime objects and ISO date '%Y-%m-%d' string are supported.
-        :type date_to: date, str, optional
         :param currency: Currency for which you need to know the rate.
             Strings like ISO numeric code, ISO code or code the Bank of Russia or Currency instance are supported.
-        :type currency: int, str, Currency:
 
         .. note:: If all arguments are specified, get the rate dynamics between the passed dates.
             If currency and one date are passed, get the rate of the specified currency at the date.
@@ -403,7 +241,7 @@ class ExchangeRateDynamics(WithRequests, FormatMixin):
 
         self.rates = rates
 
-    def __getitem__(self, item: Union[str, dt.date, dt.datetime]) -> ExchangeRate:
+    def __getitem__(self, item: Union[str, date, datetime]) -> ExchangeRate:
         """Returns the ExchangeRate by date
 
         Python date and datetime objects or '%Y-%m-%d' ISO date string are supported.
@@ -424,10 +262,10 @@ class ExchangeRateDynamics(WithRequests, FormatMixin):
         return len(self.rates)
 
     def _check_and_convert_args(self,
-                                date_from: Union[str, dt.date, dt.datetime, None],
-                                date_to: Union[str, dt.date, dt.datetime, None],
+                                date_from: Union[str, date, datetime, None],
+                                date_to: Union[str, date, datetime, None],
                                 currency: Union[str, int, Currency]
-                                ) -> Tuple[dt.datetime, dt.datetime, Currency]:
+                                ) -> Tuple[datetime, datetime, Currency]:
         """Checks arguments and converts them to proper formats from strings and integer"""
         date_from = self._datetime_from_string(date_from)
         date_to = self._datetime_from_string(date_to)
@@ -436,8 +274,8 @@ class ExchangeRateDynamics(WithRequests, FormatMixin):
             raise WrongArguments('You must specify a currency if you want to get a range of currency rates.')
 
         if not date_from and not date_to:  # if both dates are empty
-            today = dt.date.today()
-            date_from = date_to = dt.datetime(today.year, today.month, today.day)  # datetime for unification
+            today = date.today()
+            date_from = date_to = datetime(today.year, today.month, today.day)  # datetime for unification
         elif bool(date_from) ^ bool(date_to):  # if any of the dates, but not both (XOR)
             date_from = date_to = next(filter(lambda x: x, (date_from, date_to)))
 
@@ -463,7 +301,7 @@ class ExchangeRateDynamics(WithRequests, FormatMixin):
 
         return raw_data
 
-    def _parse(self, data: bytes) -> Dict[dt.datetime, ExchangeRate]:
+    def _parse(self, data: bytes) -> Dict[datetime, ExchangeRate]:
         """Parse raw XML strings with rate dynamics to the dict of ExchangeRate"""
         LOG.debug('Parsing data ...')
 
@@ -477,14 +315,14 @@ class ExchangeRateDynamics(WithRequests, FormatMixin):
             for param in child:
                 params[param.tag] = param.text
 
-            date_received = dt.datetime.strptime(child.attrib['Date'], '%d.%m.%Y')
+            date_received = datetime.strptime(child.attrib['Date'], '%d.%m.%Y')
             par = Decimal(params['Nominal'])
             value = Decimal(params['Value'].replace(',', '.'))
 
             result[date_received] = ExchangeRate(
                 currency=currency,
                 name=currency.name_eng,
-                date=date_received,
+                on_date=date_received,
                 par=par,
                 value=value,
                 rate=value / par,
